[ add ] Separate components in component folders.
