+++ src/components/Title.tsx
@@ -1,0 +1,11 @@
+export const Title = () => {
+  return (
+    <>
+      <section className="text-center mb-8">
+        <h1 className="text-3xl md:text-4xl lg:text-5xl font-bold text-gray-800">
+          Task Battle
+        </h1>
+      </section>
+    </>
+  );
+};
