+++ src/main.tsx
@@ -1,10 +1,10 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "./index.css";
-import { TaskBattle } from "./TaskBattle";
+import { App } from "./components/App";
+import "./styles/index.css";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <TaskBattle />
-  </StrictMode>,
+    <App />
+  </StrictMode>
 );
