+++ src/types/Score.ts
@@ -1,0 +1,5 @@
+export type Score = {
+  score: number;
+  label: string;
+  bgColor: string;
+};
