+++ src/types/Task.ts
@@ -1,0 +1,4 @@
+export type Task = {
+  id: number;
+  title: string;
+};
